EA-2560 delete deprected Entity

+++ entity-management-web/src/main/java/eu/europeana/entitymanagement/web/EntityAdminController.java
@@ -8,6 +8,7 @@ import eu.europeana.api.commons.web.http.HttpHeaders;
 import eu.europeana.api.commons.web.model.vocabulary.Operations;
 import eu.europeana.entitymanagement.common.config.EntityManagementConfiguration;
 import eu.europeana.entitymanagement.definitions.model.EntityRecord;
+import eu.europeana.entitymanagement.exception.EntityNotFoundException;
 import eu.europeana.entitymanagement.vocabulary.EntityProfile;
 import eu.europeana.entitymanagement.vocabulary.FormatTypes;
 import eu.europeana.entitymanagement.vocabulary.WebEntityConstants;
@@ -25,6 +26,7 @@ import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpServletRequest;
+import java.util.Optional;
 
 @RestController
 @Validated
@@ -59,7 +61,13 @@ public class EntityAdminController extends BaseRest {
         if (emConfig.isAuthEnabled()) {
             verifyWriteAccess(Operations.DELETE, request);
         }
-        EntityRecord entityRecord = entityRecordService.retrieveEntityRecord(type, identifier.toLowerCase());
+        String entityUri = EntityRecordUtils.buildEntityIdUri(type, identifier);
+        Optional<EntityRecord> entityRecordOptional = entityRecordService.retrieveEntityRecordByUri(entityUri);
+        if (entityRecordOptional.isEmpty()) {
+            throw new EntityNotFoundException(entityUri);
+        }
+        EntityRecord entityRecord = entityRecordOptional.get();
+
         LOG.debug("Deleting permanently entity : {}/{}", type, identifier);
         entityRecordService.delete(entityRecord.getEntityId());
         return ResponseEntity.noContent().build();
